+++ CHANGELOG.rst
@@ -22,6 +22,11 @@ Fixed
 - Raise exception if lazy_instance called with invalid lazy_kwargs.
 - Only add subclass defaults on defaults merging `#103
   <https://github.com/omni-us/jsonargparse/issues/103>`__.
+- Strict type and required only on final config check `#31
+  <https://github.com/omni-us/jsonargparse/issues/31>`__.
+- instantiate_classes failing for type hints with ``nargs='+'``.
+- Useful error message when init_args value invalid.
+- Specific error message when subclass dict has unexpected keys.
 
 
 v4.0.3 (2021-11-23)